Change Controller Structure

+++ src/main/java/com/example/demo/model/Medicine.java
@@ -9,14 +9,14 @@ public class Medicine {
     public Medicine() {
     }
 
-    public Medicine(int id, String name, Double price) {
+    public Medicine(Long id, String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
     @Id
-    int id;
+    Long id;
 
     @Column(name = "name")
     String name;
@@ -24,11 +24,11 @@ public class Medicine {
     @Column(name = "price")
     Double price;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
